fix message serialization
