+++ tvx/chain/types/message.go
@@ -2,8 +2,6 @@ package types
 
 import (
 	"bytes"
-	"fmt"
-	"io"
 
 	"github.com/filecoin-project/go-address"
 	"github.com/filecoin-project/specs-actors/actors/abi"
@@ -12,7 +10,6 @@ import (
 	block "github.com/ipfs/go-block-format"
 	"github.com/ipfs/go-cid"
 	"github.com/multiformats/go-multihash"
-	cbg "github.com/whyrusleeping/cbor-gen"
 )
 
 type Message struct {
@@ -39,196 +36,6 @@ type Message struct {
 
 // Below Marshalers were taken from lotus
 // https://github.com/filecoin-project/lotus/blob/95790c69e1aa568506dd39b7eeb921f7c7d6f184/chain/types/cbor_gen.go#L636-L882
-
-func (t *Message) MarshalCBOR(w io.Writer) error {
-	if t == nil {
-		_, err := w.Write(cbg.CborNull)
-		return err
-	}
-	if _, err := w.Write([]byte{136}); err != nil {
-		return err
-	}
-
-	// t.To (address.Address) (struct)
-	if err := t.To.MarshalCBOR(w); err != nil {
-		return err
-	}
-
-	// t.From (address.Address) (struct)
-	if err := t.From.MarshalCBOR(w); err != nil {
-		return err
-	}
-
-	// t.Nonce (uint64) (uint64)
-
-	if _, err := w.Write(cbg.CborEncodeMajorType(cbg.MajUnsignedInt, uint64(t.Nonce))); err != nil {
-		return err
-	}
-
-	// t.Value (big.Int) (struct)
-	if err := t.Value.MarshalCBOR(w); err != nil {
-		return err
-	}
-
-	// t.GasPrice (big.Int) (struct)
-	if err := t.GasPrice.MarshalCBOR(w); err != nil {
-		return err
-	}
-
-	// t.GasLimit (int64) (int64)
-	if t.GasLimit >= 0 {
-		if _, err := w.Write(cbg.CborEncodeMajorType(cbg.MajUnsignedInt, uint64(t.GasLimit))); err != nil {
-			return err
-		}
-	} else {
-		if _, err := w.Write(cbg.CborEncodeMajorType(cbg.MajNegativeInt, uint64(-t.GasLimit)-1)); err != nil {
-			return err
-		}
-	}
-
-	// t.Method (abi.MethodNum) (uint64)
-
-	if _, err := w.Write(cbg.CborEncodeMajorType(cbg.MajUnsignedInt, uint64(t.Method))); err != nil {
-		return err
-	}
-
-	// t.Params ([]uint8) (slice)
-	if len(t.Params) > cbg.ByteArrayMaxLen {
-		return fmt.Errorf("Byte array in field t.Params was too long")
-	}
-
-	if _, err := w.Write(cbg.CborEncodeMajorType(cbg.MajByteString, uint64(len(t.Params)))); err != nil {
-		return err
-	}
-	if _, err := w.Write(t.Params); err != nil {
-		return err
-	}
-	return nil
-}
-
-func (t *Message) UnmarshalCBOR(r io.Reader) error {
-	br := cbg.GetPeeker(r)
-
-	maj, extra, err := cbg.CborReadHeader(br)
-	if err != nil {
-		return err
-	}
-	if maj != cbg.MajArray {
-		return fmt.Errorf("cbor input should be of type array")
-	}
-
-	if extra != 8 {
-		return fmt.Errorf("cbor input had wrong number of fields")
-	}
-
-	// t.To (address.Address) (struct)
-
-	{
-
-		if err := t.To.UnmarshalCBOR(br); err != nil {
-			return fmt.Errorf("unmarshaling t.To: %w", err)
-		}
-
-	}
-	// t.From (address.Address) (struct)
-
-	{
-
-		if err := t.From.UnmarshalCBOR(br); err != nil {
-			return fmt.Errorf("unmarshaling t.From: %w", err)
-		}
-
-	}
-	// t.Nonce (uint64) (uint64)
-
-	{
-
-		maj, extra, err = cbg.CborReadHeader(br)
-		if err != nil {
-			return err
-		}
-		if maj != cbg.MajUnsignedInt {
-			return fmt.Errorf("wrong type for uint64 field")
-		}
-		t.Nonce = uint64(extra)
-
-	}
-	// t.Value (big.Int) (struct)
-
-	{
-
-		if err := t.Value.UnmarshalCBOR(br); err != nil {
-			return fmt.Errorf("unmarshaling t.Value: %w", err)
-		}
-
-	}
-	// t.GasPrice (big.Int) (struct)
-
-	{
-
-		if err := t.GasPrice.UnmarshalCBOR(br); err != nil {
-			return fmt.Errorf("unmarshaling t.GasPrice: %w", err)
-		}
-
-	}
-	// t.GasLimit (int64) (int64)
-	{
-		maj, extra, err := cbg.CborReadHeader(br)
-		var extraI int64
-		if err != nil {
-			return err
-		}
-		switch maj {
-		case cbg.MajUnsignedInt:
-			extraI = int64(extra)
-			if extraI < 0 {
-				return fmt.Errorf("int64 positive overflow")
-			}
-		case cbg.MajNegativeInt:
-			extraI = int64(extra)
-			if extraI < 0 {
-				return fmt.Errorf("int64 negative oveflow")
-			}
-			extraI = -1 - extraI
-		default:
-			return fmt.Errorf("wrong type for int64 field: %d", maj)
-		}
-
-		t.GasLimit = int64(extraI)
-	}
-	// t.Method (abi.MethodNum) (uint64)
-
-	{
-
-		maj, extra, err = cbg.CborReadHeader(br)
-		if err != nil {
-			return err
-		}
-		if maj != cbg.MajUnsignedInt {
-			return fmt.Errorf("wrong type for uint64 field")
-		}
-		t.Method = abi.MethodNum(extra)
-
-	}
-	// t.Params ([]uint8) (slice)
-
-	maj, extra, err = cbg.CborReadHeader(br)
-	if err != nil {
-		return err
-	}
-
-	if extra > cbg.ByteArrayMaxLen {
-		return fmt.Errorf("t.Params: byte array too large (%d)", extra)
-	}
-	if maj != cbg.MajByteString {
-		return fmt.Errorf("expected byte array")
-	}
-	t.Params = make([]byte, extra)
-	if _, err := io.ReadFull(br, t.Params); err != nil {
-		return err
-	}
-	return nil
-}
 
 func (m *Message) MustSerialize() []byte {
 	buf := new(bytes.Buffer)
@@ -269,63 +76,6 @@ func (m *Message) Cid() cid.Cid {
 type SignedMessage struct {
 	Message   Message
 	Signature crypto.Signature
-}
-
-func (t *SignedMessage) MarshalCBOR(w io.Writer) error {
-	if t == nil {
-		_, err := w.Write(cbg.CborNull)
-		return err
-	}
-	if _, err := w.Write([]byte{130}); err != nil {
-		return err
-	}
-
-	// t.Message (types.Message) (struct)
-	if err := t.Message.MarshalCBOR(w); err != nil {
-		return err
-	}
-
-	// t.Signature (crypto.Signature) (struct)
-	if err := t.Signature.MarshalCBOR(w); err != nil {
-		return err
-	}
-	return nil
-}
-
-func (t *SignedMessage) UnmarshalCBOR(r io.Reader) error {
-	br := cbg.GetPeeker(r)
-
-	maj, extra, err := cbg.CborReadHeader(br)
-	if err != nil {
-		return err
-	}
-	if maj != cbg.MajArray {
-		return fmt.Errorf("cbor input should be of type array")
-	}
-
-	if extra != 2 {
-		return fmt.Errorf("cbor input had wrong number of fields")
-	}
-
-	// t.Message (types.Message) (struct)
-
-	{
-
-		if err := t.Message.UnmarshalCBOR(br); err != nil {
-			return fmt.Errorf("unmarshaling t.Message: %w", err)
-		}
-
-	}
-	// t.Signature (crypto.Signature) (struct)
-
-	{
-
-		if err := t.Signature.UnmarshalCBOR(br); err != nil {
-			return fmt.Errorf("unmarshaling t.Signature: %w", err)
-		}
-
-	}
-	return nil
 }
 
 func (sm *SignedMessage) Serialize() ([]byte, error) {
